Consolidate test/example and test/upgrade into suite-config
