+++ test/default.env
@@ -27,7 +27,5 @@ MAX_JOBS=4
 
 # TEST RUNNER CONFIG
 # options: "trace", "debug", "info", "warn", "error", "silent"	
-DOCKER_WIKIBASE_REPO_NAME=test-wikibase-1
-DOCKER_MYSQL_NAME=test-mysql-1
-DOCKER_WIKIBASE_CLIENT_NAME=wikibase-client-1
 SELENIUM_LOG_LEVEL=error
+DOCKER_MYSQL_NAME=test-mysql-1
