--- test/docker-compose.example.yml
@@ -1,31 +1,0 @@
-# Wikibase Suite from Example directory overrides
-version: '3.4'
-
-services:
-
-  wikibase-selenium-test:
-    environment:
-      MW_SERVER: "http://${WIKIBASE_HOST}"
-      QS_SERVER: "http://${QUICKSTATEMENTS_HOST}"
-      WDQS_FRONTEND_SERVER: "${WDQS_FRONTEND_HOST}"
-      WDQS_SERVER: "wdqs.svc:9999"
-      WDQS_PROXY_SERVER: "http://wdqs-proxy.svc"
-      MW_SCRIPT_PATH: "/w"
-
-  wikibase:
-    environment:
-      - QS_PUBLIC_SCHEME_HOST_AND_PORT=http://quickstatements.svc:80
-    volumes:
-      # this gets executed in example/
-      - ./suite-config/${SUITE_CONFIG_NAME}/LocalSettings.php:/var/www/html/LocalSettings.d/LocalSettings.test.php
-      - ./log/${SUITE_CONFIG_NAME}/wikibase:/var/log/mediawiki
-      - ../example/extra-install.sh:/extra-install.sh
-
-  wikibase-jobrunner:
-    volumes:
-      - ../example/jobrunner-entrypoint.sh:/jobrunner-entrypoint.sh
-
-  quickstatements:
-    environment:
-      - QS_PUBLIC_SCHEME_HOST_AND_PORT=http://quickstatements.svc:80
-
