--- test/example.env
@@ -1,4 +1,0 @@
-DOCKER_WIKIBASE_REPO_NAME=example-wikibase-1
-DOCKER_MYSQL_NAME=example-mysql-1
-
-MW_WG_ENABLE_UPLOADS=true
