--- test/setup.sh
@@ -1,1 +1,0 @@
-#!/bin/bash
