+++ test/suite-config/example/LocalSettings.php
@@ -1,3 +1,4 @@
 <?php
+
 // EntitySchema
 $wgEntitySchemaShExSimpleUrl = "http://validator.svc";
