+++ test/suite-config/example/example.env
@@ -1,0 +1,8 @@
+DOCKER_MYSQL_NAME=example-mysql-1
+MW_WG_ENABLE_UPLOADS=true
+MW_SERVER="http://${WIKIBASE_HOST}"
+QS_SERVER="http://${QUICKSTATEMENTS_HOST}"
+WDQS_FRONTEND_SERVER="${WDQS_FRONTEND_HOST}"
+WDQS_SERVER="wdqs.svc:9999"
+WDQS_PROXY_SERVER="http://wdqs-proxy.svc"
+MW_SCRIPT_PATH="/w"
