+++ test/suite-config/wikibase-create-db.sql
